Rm functions related to Old AddBatch
